Feature: populated migrations files and generated users and blogposts tables with model changes
