--- .coveragerc
@@ -1,2 +1,0 @@
- [run]
- omit = src/tests/*
