--- src/__init__.py
@@ -1,7 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-
-# For initializing our db
-db = SQLAlchemy()
-
-bcrypt = Bcrypt()
