+++ src/config.py
@@ -7,8 +7,8 @@ class Development(object):
     """
     DEBUG = True
     TESTING = False
-    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
-    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL')
+    JWT_SECRET_KEY = 'hhgaghhgsdhdhdd'
+    SQLALCHEMY_DATABASE_URI = 'postgres://divyanshuhome:password@localhost:5432/blog_api_db'
 
 
 class Production(object):
@@ -21,18 +21,18 @@ class Production(object):
     JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
 
 
-class Testing(object):
-    """
-    Development environment configuration
-    """
-    TESTING = True
-    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
-    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_TEST_URL')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-
+# class Testing(object):
+#     """
+#     Development environment configuration
+#     """
+#     TESTING = True
+#     JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
+#     SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_TEST_URL')
+#     SQLALCHEMY_TRACK_MODIFICATIONS = False
+#
+#
 app_config = {
     'development': Development,
-    'production': Production,
-    'testing': Testing
+    'production': Production
+    #     'testing': Testing
 }
