+++ src/model_blog/__init__.py
@@ -1,0 +1,9 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+# from .BlogPostModel import BlogPostModel, BlogpostSchema
+# from .UserModel import UserModel, UserSchema
+
+# initialize our db
+db = SQLAlchemy()
+bcrypt = Bcrypt()
+
